Simplify bin/deploy
- Use the shebang line for POSIX sh.
- Default the argument to 'staging'
- Remove `set -e`.

The `set -e` functionality is not a 100% guarantee of expected semantics
(this is explored more by David Röthlisberger in [Shell scripts: Perils
of “set -e” for error handling][set-e]). In this case, the fact that the
script is two lines means that we can simply add a `&&` to the end of
the first line to get the desired semantics. Not using `set -e` means
that we have to understand how various commands can fail which,
honestly, is a feature.

[set-e]: http://david.rothlis.net/shell-set-e/